・敵の挙動の改善 プレイヤーを目で見る 攻撃範囲の詳細化 予測射撃
・Vector2クラスの実装
入れた方が便利な気がしたので

・コンポーネントの実装
同じような動きをする敵がいた場合それらを一つの関数などで入れられる方がよりゲーム制作がしやすくなると思ったため実装
名前はこれでいいか迷っているので変更の可能性はあり

+++ TR1_Dxlib/Global.h
@@ -4,6 +4,7 @@
 #include "Struct.h"
 #include "enum.h"
 #include "math.h"
+#include "Vector2.h"
 
 // ウィンドウのタイトルに表示する文字列
 const char TITLE[] = "TR1";
@@ -15,10 +16,10 @@ const int WIN_WIDTH = 640;
 const int WIN_HEIGHT = 480;
 
 //円周率
-const double PI = 3.141592;
+const float PI = 3.141592f;
 
 //重力
-const double G = 9.8;
+const float G = 9.8f;
 
 //マップチップのサイズ
 const int MAP_SIZE = 32;
